Separate and rename Docker image scripts

+++ michigan-private/docker_reset.sh
@@ -1,11 +1,21 @@
 #!/bin/bash
+#
+# Resets the Docker image. Run this when you need to start with a fresh Docker image (e.g., reset any installations
+# you did inside the image).
 
 SCRIPT_DIR="$(cd $(dirname $0); pwd)"
 PROJ_DIR="$(cd $SCRIPT_DIR/..; pwd)"
 
 cd "$PROJ_DIR"
-docker container stop spider_test
-docker container rm spider_test
+# Stop Docker image if it's running
+if [ -n "$(docker ps --filter 'name=spider_test' -q)" ]; then
+  docker container stop spider_test
+fi
+# Delete Docker image if it exists
+if [ -n "$(docker ps -a | grep spider_test)" ]; then
+  docker container rm spider_test
+fi
+
 docker run -d \
   -it \
   --name spider_test \
@@ -19,5 +29,3 @@ docker run -d \
   --env "GIT_AUTHOR_NAME=$(git config user.name)" \
   --env "GIT_AUTHOR_EMAIL=$(git config user.email)" \
   spider
-
-docker exec -it spider_test bash
